wsgi/webconfgen/requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
binaryornot==0.3.0
celery==3.1.18
Django==1.8.3
djangorestframework==3.1.3
kombu==3.0.26
librabbitmq==1.6.1
psycopg2==2.6.1
pytz==2015.4
redis==2.10.3
uuid==1.30
wheel==0.24.0
